TIP-856: Move Pim\Analytics classes resources to Akeneo\Tool
