+++ phpspec.yml.dist
@@ -5,10 +5,10 @@ extensions:
 suites:
     ## Akeneo Components
     Analytics:
-        namespace: Akeneo\Component\Analytics
-        psr4_prefix: Akeneo\Component\Analytics
-        spec_path: src/Akeneo/Component/Analytics
-        src_path: src/Akeneo/Component/Analytics
+        namespace: Akeneo\Tool\Component\Analytics
+        psr4_prefix: Akeneo\Tool\Component\Analytics
+        spec_path: src/Akeneo/Tool/Component/Analytics
+        src_path: src/Akeneo/Tool/Component/Analytics
     Batch:
         namespace: Akeneo\Tool\Component\Batch
         psr4_prefix: Akeneo\Tool\Component\Batch
