+++ src/Akeneo/Tool/Component/Analytics/DataCollectorInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Component\Analytics;
+namespace Akeneo\Tool\Component\Analytics;
 
 /**
  * Interface DataCollectorInterface
