+++ src/Pim/Bundle/AnalyticsBundle/DataCollector/DBDataCollector.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\AnalyticsBundle\DataCollector;
 
-use Akeneo\Component\Analytics\DataCollectorInterface;
+use Akeneo\Tool\Component\Analytics\DataCollectorInterface;
 use Akeneo\Tool\Component\StorageUtils\Repository\CountableRepositoryInterface;
 use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
