+++ src/Pim/Bundle/AnalyticsBundle/DataCollector/StorageDataCollector.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\AnalyticsBundle\DataCollector;
 
-use Akeneo\Component\Analytics\DataCollectorInterface;
+use Akeneo\Tool\Component\Analytics\DataCollectorInterface;
 
 /**
  * Collects MySQL version.
