+++ src/Pim/Bundle/AnalyticsBundle/spec/DataCollector/AttributeDataCollectorSpec.php
@@ -2,7 +2,7 @@
 
 namespace spec\Pim\Bundle\AnalyticsBundle\DataCollector;
 
-use Akeneo\Component\Analytics\DataCollectorInterface;
+use Akeneo\Tool\Component\Analytics\DataCollectorInterface;
 use Akeneo\Tool\Component\StorageUtils\Repository\CountableRepositoryInterface;
 use Pim\Bundle\AnalyticsBundle\DataCollector\AttributeDataCollector;
 use PhpSpec\ObjectBehavior;
